fix staticcasts

+++ src/map/atcommand.cpp
@@ -2701,7 +2701,7 @@ ACMD_FUNC(memo)
 		int32 i;
 		clif_displaymessage(sd->fd,  msg_txt(sd,668)); // Your actual memo positions are:
 
-		for( i = 0; i < MAX_MEMOPOINTS + extendedMemo; i++ )
+		for( i = 0; i < static_cast<int32>( MAX_MEMOPOINTS + extendedMemo ); i++ )
 		{
 			if( strcmp( "", sd->status.memo_point[i].map ) != 0 )
 				sprintf( atcmd_output, "%d - %s (%d,%d)", i, sd->status.memo_point[i].map, sd->status.memo_point[i].x, sd->status.memo_point[i].y );
@@ -2712,9 +2712,9 @@ ACMD_FUNC(memo)
 		return 0;
  	}
 
-	if( position < 0 || position >= MAX_MEMOPOINTS + extendedMemo)
+	if( position < 0 || position >= static_cast<int32>( MAX_MEMOPOINTS + extendedMemo ) )
 	{
-		sprintf(atcmd_output, msg_txt(sd,1008), 0, MAX_MEMOPOINTS+ extendedMemo -1); // Please enter a valid position (usage: @memo <memo_position:%d-%d>).
+		sprintf(atcmd_output, msg_txt(sd,1008), 0, static_cast<int32>( MAX_MEMOPOINTS + extendedMemo - 1 ) ); // Please enter a valid position (usage: @memo <memo_position:%d-%d>).
 		clif_displaymessage(fd, atcmd_output);
 		return -1;
 	}
